Move traits into their own modules
This makes some imports slightly more intuitive. `lib.rs` still re-exports
them.
